refactor: automated socket listening and mjs configs
