+++ src/constants.js
@@ -1,0 +1,9 @@
+const SOCKET = {
+    prod: "https://melted-patience-leindfraust.koyeb.app/",
+    preview: "https://zefer-socket.onrender.com/",
+    dev: "http://localhost:5000",
+};
+
+module.exports = {
+    SOCKET,
+};
